deploy/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pillow==10.4.0

torch==2.3.1
torchvision==0.18.1
open_clip_torch==2.26.1
google-cloud-storage==2.17.0